add companycontroller delete safeguarding

+++ app/Http/Controllers/CompanyController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Events\PermissionDenied;
 use App\Models\Company;
 use Illuminate\Http\Request;
 
@@ -24,6 +25,11 @@ class CompanyController extends Controller
 
     public function destroy(Company $company)
     {
+        if(!auth()->user()->can('companies.delete')){
+            event(new PermissionDenied());
+            return back();
+        }
+
         $company->delete();
         return back();
     }
